svd: full device parsing
This is a huge expansion of svada to support a lot of new features
not previously supported.
The main motivation behind these changes is to parse the whole device
in a way (more or less) following the SVD specification, and exposing
more of the content in the SVD file, to be able to use svada
in a wider set of use cases.

The API for parsing an SVD file has been changed from
"parse_peripheral()" to just "parse()", and now returns a
Device instance instead of a Peripheral.
This Device class represents the complete SVD device and mirrors the
root element in the SVD file.

The Peripheral implementation was changed so that the Device/Peripheral
no longer deals directly with the XML tree, instead accessing the SVD
properties through an intermediary "binding" layer.
This binding layer uses the lxml objectify API that maps XML elements
to Python objects. Custom classes were written on top of the
lxml/objectify API to make accessing the XML elements/attributes
convenient and to minimize the boilerplate associated with reading
these properties.
The goal for this binding API was to make a Python interface that
abstracts the whole SVD device, exposing every XML element and
attribute as a typed Python class or attribute.

The benefit of separating out the low-level XML interpretation is that
the Device/Peripheral now only needs to deal with high level SVD
semantics.

The existing Peripheral API was changed a lot, with the intentions of

* exposing more of the SVD contents
* better match the structure of the SVD contents
* offer a more convenient API for accessing the new features

Registers are no longer expanded/flattened, and are instead nested
similarly to how they are layed out in the SVD file.
Each register level element is represented using a class specific
to the type of element, like an Array, Struct or Register.
Register names now also (mostly) match the names in the SVD file
exactly (except for register names containing a "[%s]" suffix).

The register classes are divided into regular and "flat" registers.
The difference between these types is that regular register structures
use the dimensions defined on SVD cluster and register elements to
actually expand registers into an array representation, while flat
registers correspond directly to the element layout in the SVD file,
without any dimensions.
Regular registers also have content stored in them, whereas flat
registers do not.

These two types of registers are meant to support two different use
cases.
The regular registers work similarly to C structures, and can be used
to keep track of the memory of a peripheral.
The flat registers are intended for users that simply want to read
the properties defined on the SVD elements and inspect the structure
of the SVD device.

Registers can be accessed from a Peripheral or from a register class
by their path relative to the element they are acccessed from.
A register path is either a C-expression like string such as
"REGA.REGB[123]", an equivalent sequence of names and indices like
("REGA", "REGB", 123), or a mix of both.
The EPath and FEPath classes were introduced to handle register paths
in a structured and well-defined way, and can be used to refer to
regular and flat registers respectively.

Peripheral memory is now handled by the MemoryBlock class.
This class uses numpy arrays to handle data types and byteorder in a
well defined way, and also hande the content of large peripherals more
efficiently (note that it can still take some time when initializing the
content of large arrays).

The Device/Peripheral API now uses a custom set of errors with
semantics specific to SVD.

Ref: NCSDK-21485

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>