svd: add helper for populating and outputting device memory
Adds a utility module with helpers for populating and outputting
device memory.

The following formats are supported for input/output:
* A dict mapping address -> value (at a byte level)
* A dict mapping a device-like structure to register/field values

The following format is currently only supported for output:
* A bytearray containing device memory

The APIs for generating outputs offer a flexible "selector"
mechanism that lets the user decide which part of the device to
output.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>