device: Fix off-by-one error in Register.set_content
This method does a bit width check to reject values that wouldn't fit in
a given register. This used to be done using a floating-point expression
which gave wrong results for powers of 2. The correct formula would be:

    math.floor(math.log2(x)) + 1

That said, it's way more efficient to use `int.bit_length` instead.

Note: currently, this fix will not be noticeable to the user, because
any register overflows missed by svada would've been caught by numpy.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>